+++ __init__.py
@@ -1,2 +1,0 @@
-# for testing only!
-from .lib_csv import *