+++ build_docs.py
@@ -25,9 +25,8 @@ from docopt import docopt       # type: ignore
 import lib_log_utils            # type: ignore
 import rst_include              # type: ignore
 
-
-# CONSTANTS & PROJECT SPECIFIC FUNCTIONS
-codeclimate_link_hash = "7fa21a0ced3820c5faa9"  # for lib_csv
+# PROJ
+import project_conf
 
 
 def project_specific(repository_slug: str, repository: str, repository_dashed: str) -> None:
@@ -57,28 +56,12 @@ def main(args: Dict[str, str]) -> None:
                         target='./README.rst')
 
     logger.info('replace repository related strings')
-    rst_include.rst_str_replace(source='./README.rst',
-                                target='',
-                                old='{repository_slug}',
-                                new=travis_repo_slug,
-                                inplace=True)
-    rst_include.rst_str_replace(source='./README.rst',
-                                target='',
-                                old='{repository}',
-                                new=repository,
-                                inplace=True)
-    rst_include.rst_str_replace(source='./README.rst',
-                                target='',
-                                old='{repository_dashed}',
-                                new=repository_dashed,
-                                inplace=True)
+    rst_include.rst_str_replace(source='./README.rst', target='', old='{repository_slug}', new=travis_repo_slug, inplace=True)
+    rst_include.rst_str_replace(source='./README.rst', target='', old='{repository}', new=repository, inplace=True)
+    rst_include.rst_str_replace(source='./README.rst', target='', old='{double_underline_repository}', new='=' * len(repository), inplace=True)
+    rst_include.rst_str_replace(source='./README.rst', target='', old='{repository_dashed}', new=repository_dashed, inplace=True)
     rst_include.rst_str_replace(source='./README.rst', target='', old='{last_update_yyyy}', new=str(datetime.date.today().year + 1), inplace=True)
-    rst_include.rst_str_replace(source='./README.rst',
-                                target='',
-                                old='{codeclimate_link_hash}',
-                                new=codeclimate_link_hash,
-                                inplace=True)
-
+    rst_include.rst_str_replace(source='./README.rst', target='', old='{codeclimate_link_hash}', new=project_conf.codeclimate_link_hash, inplace=True)
     logger.info('done')
     sys.exit(0)