+++ CHANGES.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+0.1.0
+-----
+2020-05-27:
+    - new build matrix
+    - mypy strict type testing
+    - fix title in pypi documentation
+    - drop python2.7 - python 3.4 support
+
 0.0.1
 -----
-2020-05-06: Initial public release
+2020-05-06:
+    - Initial public release